conf

+++ entrypoint.sh
@@ -1,3 +1,8 @@
 #!/bin/sh
-cp -f ${WEB_CONF} /etc/nginx/conf.d/default.conf
+
+if [ -f "${WEB_CONF}" ];
+then
+	cp -f ${WEB_CONF} /etc/nginx/conf.d/default.conf
+fi
+
 /usr/sbin/nginx -g "daemon off;"
